requirements.txt
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
numpy==2.2.6
pillow==12.0.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.1
